+++ ClientSide/src/Interface/Ui/RunRulesPage/Questions/QuestionItems/QuestionDoubleType.js
@@ -7,14 +7,13 @@ import PropTypes from 'prop-types';
 
 export default class QuestionDoubleType extends React.Component {
     constructor(props) {
-        super(props);
-        this.state ={
-            answered: false,
-        }
+      super(props);
     }
 
 
-    
+    state ={
+        answered: false,
+    }
 
     // prop types and default values
     static propTypes = {
@@ -44,13 +43,6 @@ export default class QuestionDoubleType extends React.Component {
     _onCancel=()=>{
         this.setState({inputValue:{}, answered: !this.state.answered});
     }
-
-    _onEditAnswer=()=>{
-        debugger;
-        if(this.props.onEditAnswer){
-            this.props.onEditAnswer(this.props.question);
-        }
-    }
      
     // component render method
     render() {
@@ -64,7 +56,6 @@ export default class QuestionDoubleType extends React.Component {
                             <Icon name='sort numeric ascending' />
                             <input />
                         </Input>
-                        <Button color='yellow' floated='right' onClick={this._onEditAnswer}>Edit</Button>
                     </Segment>   
                 :
                     <Segment attached='bottom'>
