+++ ClientSide/src/Interface/Ui/RunRulesPage/Questions/QuestionItems/QuestionStringType/QuestionStringType.js
@@ -9,13 +9,14 @@ import './QuestionStringType.scss';
 
 export default class QuestionStringType extends React.Component {
     constructor(props) {
-        super(props);
-        this.state ={
-            answered: false,
-            inputValue:'',
-        }
+      super(props);
     }
 
+
+    state ={
+        answered: false,
+        inputValue:'',
+    }
 
     // prop types and default values
     static propTypes = {
@@ -37,13 +38,6 @@ export default class QuestionStringType extends React.Component {
     _onCancel=()=>{
         this.setState({inputValue:{}, answered: !this.state.answered});
     }
-
-    _onEditAnswer=()=>{
-        debugger;
-        if(this.props.onEditAnswer){
-            this.props.onEditAnswer(this.state.question);
-        }
-    }
      
     // component render method
     render() {
@@ -54,7 +48,6 @@ export default class QuestionStringType extends React.Component {
                 {this.state.answered?
                     <Message className='stringType-bottom-message' attached='bottom' color='olive'>
                         <Header floated='right' size = 'large'>{this.state.inputValue}</Header>
-                        <Button color='yellow' floated='right' onClick={this._onEditAnswer}>Edit</Button>                
                     </Message>
                     :
                     <Segment attached='bottom'>
