+++ ClientSide/src/Interface/Ui/RunRulesPage/Questions/QuestionItems/QuestionUrlType.js
@@ -6,14 +6,16 @@ import PropTypes from 'prop-types';
 
 
 
-export default class QuestionUuidType extends React.Component {
+export default class QuestionUrlType extends React.Component {
     constructor(props) {
       super(props);
-      this.state = {
-        answered: false,
-      }
     }
 
+
+
+    state ={
+        answered: false,
+    }
 
     // prop types and default values
     static propTypes = {
@@ -22,7 +24,7 @@ export default class QuestionUuidType extends React.Component {
 
     _onChange=(event, newValue)=>{
         
-        let regEx = /^[0-9a-f]{8}-([0-9a-f]{4}-){3}[0-9a-f]{12}/;
+        let regEx = /^(ht|f)tps?\\:(\\p{Graph}|\\p{XDigit}|\\p{Space})*$/;
         if(!regEx.test(newValue))
         {
             this.setState({inputError: !this.state.inputError});
@@ -35,19 +37,13 @@ export default class QuestionUuidType extends React.Component {
 
     _onSave=()=>{
         if(this.props.onSave){
-            this.props.onSave(this.props.question, {answer:this.state.inputValue, type:'uuid'});
+            this.props.onSave(this.props.question, {answer:this.state.inputValue, type:'url'});
             this.setState({answered: !this.state.answered});
         }
     }
 
     _onCancel=()=>{
         this.setState({inputValue:{}, answered: !this.state.answered});
-    }
-
-    _onEditAnswer=()=>{
-        if(this.props.onEditAnswer){
-            this.props.onEditAnswer(this.state.question);
-        }
     }
      
     // component render method
@@ -56,18 +52,17 @@ export default class QuestionUuidType extends React.Component {
         return (
             <Segment.Group raised className='questionIntItem'>
                 <Message attached='top' info header= {question}/>
-                
                 {this.state.answered?
                     <Segment attached='bottom' inverted color='green'>
-                         <Message className='uuidType-bottom-message' attached='bottom' color='olive'>
-                            <Header floated='right' size = 'large'>{this.state.inputValue}</Header>
-                            <Button color='yellow' floated='right' onClick={this._onEditAnswer}>Edit</Button>
-                        </Message>
+                        <Input iconPosition='left' placeholder='Please enter alphabetical characters' value={this.state.inputValue}>
+                            <Icon name='sort alphabet ascending' />
+                            <input />
+                        </Input>
                     </Segment>
                     :
                     this.state.inputError?
                         <Segment attached='bottom'>
-                            <Input error iconPosition='left' placeholder='Please enter numerical values' value={this.state.inputValue} onChange={this._onChange}>
+                            <Input error iconPosition='left' placeholder='Please enter URL' onChange={this._onChange}>
                                 <Icon name='sort alphabet ascending' />
                                 <input />
                             </Input>
@@ -79,7 +74,7 @@ export default class QuestionUuidType extends React.Component {
                         </Segment>
                         :
                         <Segment attached='bottom'>
-                            <Input iconPosition='left' placeholder='Please enter numerical values' value={this.state.inputValue} onChange={this._onChange}>
+                            <Input iconPosition='left' placeholder='Please enter URL' onChange={this._onChange}>
                                 <Icon name='sort alphabet ascending' />
                                 <input />
                             </Input>
@@ -89,9 +84,9 @@ export default class QuestionUuidType extends React.Component {
                                 <Button negative onClick={this._onCancel}>Cancel</Button>
                             </Button.Group>
                         </Segment>}
+                        
                 
             </Segment.Group>
         );
-        
     }
 }
