+++ ClientSide/src/Interface/Ui/RunRulesPage/Questions/QuestionItems/QuestionUuidType.js
@@ -6,13 +6,15 @@ import PropTypes from 'prop-types';
 
 
 
-export default class QuestionUrlType extends React.Component {
+export default class QuestionUuidType extends React.Component {
     constructor(props) {
-        super(props);
-        this.state ={
-            answered: false,
-        }
+      super(props);
+
+      state = {
+        answered: false,
+      }
     }
+
 
     // prop types and default values
     static propTypes = {
@@ -21,7 +23,7 @@ export default class QuestionUrlType extends React.Component {
 
     _onChange=(event, newValue)=>{
         
-        let regEx = /^(ht|f)tps?\\:(\\p{Graph}|\\p{XDigit}|\\p{Space})*$/;
+        let regEx = /^[0-9a-f]{8}-([0-9a-f]{4}-){3}[0-9a-f]{12}/;
         if(!regEx.test(newValue))
         {
             this.setState({inputError: !this.state.inputError});
@@ -34,7 +36,7 @@ export default class QuestionUrlType extends React.Component {
 
     _onSave=()=>{
         if(this.props.onSave){
-            this.props.onSave(this.props.question, {answer:this.state.inputValue, type:'url'});
+            this.props.onSave(this.props.question, {answer:this.state.inputValue, type:'uuid'});
             this.setState({answered: !this.state.answered});
         }
     }
@@ -42,14 +44,6 @@ export default class QuestionUrlType extends React.Component {
     _onCancel=()=>{
         this.setState({inputValue:{}, answered: !this.state.answered});
     }
-
-    _onEditAnswer=()=>{
-        debugger;
-        if(this.props.onEditAnswer){
-            this.props.onEditAnswer(this.state.question);
-        }
-    }
-     
      
     // component render method
     render() {
@@ -57,17 +51,18 @@ export default class QuestionUrlType extends React.Component {
         return (
             <Segment.Group raised className='questionIntItem'>
                 <Message attached='top' info header= {question}/>
+                
                 {this.state.answered?
                     <Segment attached='bottom' inverted color='green'>
-                        <Message className='urlType-bottom-message' attached='bottom' color='olive'>
-                            <Header floated='right' size = 'large'>{this.state.inputValue}</Header>
-                            <Button color='yellow' floated='right' onClick={this._onEditAnswer}>Edit</Button>
-                        </Message>
+                        <Input disabled iconPosition='left' placeholder='Please enter numerical values' value={this.state.inputValue}>
+                            <Icon name='sort alphabet ascending' />
+                            <input />
+                        </Input>
                     </Segment>
                     :
                     this.state.inputError?
                         <Segment attached='bottom'>
-                            <Input error iconPosition='left' placeholder='Please enter URL' onChange={this._onChange}>
+                            <Input error iconPosition='left' placeholder='Please enter numerical values' value={this.state.inputValue} onChange={this._onChange}>
                                 <Icon name='sort alphabet ascending' />
                                 <input />
                             </Input>
@@ -79,7 +74,7 @@ export default class QuestionUrlType extends React.Component {
                         </Segment>
                         :
                         <Segment attached='bottom'>
-                            <Input iconPosition='left' placeholder='Please enter URL' onChange={this._onChange}>
+                            <Input iconPosition='left' placeholder='Please enter numerical values' value={this.state.inputValue} onChange={this._onChange}>
                                 <Icon name='sort alphabet ascending' />
                                 <input />
                             </Input>
@@ -89,9 +84,9 @@ export default class QuestionUrlType extends React.Component {
                                 <Button negative onClick={this._onCancel}>Cancel</Button>
                             </Button.Group>
                         </Segment>}
-                        
                 
             </Segment.Group>
         );
+        
     }
 }
